Implementesd Timerseries class and getFrequency().
Implemented the class for Timerseries. The comparison is preliminarily build
with a simple algorithm, that runs in quadratic time (O(n^2)) for comparisons of
two Timerseries. This will be improved in the future.

I also implemented to methods: Timer.getFrequency() and
Timestamp.getFrequency(). These two return the Frequency in Hertz resulting from
the time stored in these objects.